style_transfer/st_manifold_alignment_mast/MAST/requirements.txt
imageio>=2.6.1
imgaug>=0.2.5
opencv-python>=4.1.2
matplotlib>=3.1.1
numpy>=1.17.1
Pillow>=6.2.1
PyYAML>=5.1.2
scikit-image>=0.15.0
scikit-learn>=0.21.3
scipy==1.2.2
seaborn>=0.11.0
tensorboard>=2.3.0
torch>=1.2.0
torchfile>=0.1.0
torchvision>=0.4.0
tqdm>=4.50.2
yacs>=0.1.6
PyQt5>=5.15
